requirements.txt
pillow>=9.0.0
keyboard>=0.13.5
